Up to work on Turno UPDATE and DELETE methods

+++ src/main/java/com/dh/clinica/controller/CrudController.java
@@ -15,7 +15,7 @@ public interface CrudController<T> {
 
     ResponseEntity<?> actualizar(@RequestBody T t) throws BadRequestException, ResourceNotFoundException;
 
-    ResponseEntity<?> buscar(@PathVariable Long id) throws BadRequestException, ResourceNotFoundException;
+    ResponseEntity<?> buscar(@PathVariable Integer id) throws BadRequestException, ResourceNotFoundException;
 
-    ResponseEntity<String> eliminar(@PathVariable Long id) throws BadRequestException, ResourceNotFoundException;
+    ResponseEntity<String> eliminar(@PathVariable Integer id) throws BadRequestException, ResourceNotFoundException;
 }
