+++ src/main/java/com/dh/clinica/repository/impl/DomicilioRepository.java
@@ -3,5 +3,5 @@ package com.dh.clinica.repository.impl;
 import com.dh.clinica.entity.Domicilio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DomicilioRepository extends JpaRepository<Domicilio, Long> { // Por defecto toma ID como Long, por más que lo setee como Integer
+public interface DomicilioRepository extends JpaRepository<Domicilio, Integer> { // Por defecto toma ID como Long, por más que lo setee como Integer
 }
