+++ src/main/java/com/dh/clinica/repository/impl/OdontologoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.Optional;
 import java.util.Set;
 
-public interface OdontologoRepository extends JpaRepository<Odontologo, Long> { // Por defecto toma ID como Long, por más que lo setee como Integer
+public interface OdontologoRepository extends JpaRepository<Odontologo, Integer> { // Por defecto toma ID como Long, por más que lo setee como Integer
     // Para crear un nuevo odontólogo usamos .save()
 
     @Query(/*"SELECT o */"FROM Odontologo o WHERE o.apellido LIKE %:apellido%"/*= ?1"*/)
