+++ src/main/java/com/dh/clinica/repository/impl/TurnoRepository.java
@@ -3,5 +3,5 @@ package com.dh.clinica.repository.impl;
 import com.dh.clinica.entity.Turno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TurnoRepository extends JpaRepository<Turno, Long> {
+public interface TurnoRepository extends JpaRepository<Turno, Integer> {
 }
